+++ packages/ngx-layout/src/lib/hero/public-api.ts
@@ -1,4 +1,0 @@
-export { HeroModule } from './hero.module';
-export { HeroComponent } from './components/hero/hero.component';
-export { HeroCardDirective } from './directives/hero-card.directive';
-export { HeroCtaDirective } from './directives/hero-cta.directive';