+++ packages/ngx-layout/src/public-api.ts
@@ -1,6 +1,3 @@
 export * from './lib/footer/public-api';
 export * from './lib/header/public-api';
-export * from './lib/hero/public-api';
 export * from './lib/modal/public-api';
-export * from './lib/pane/public-api';
-export * from './lib/sidebar/public-api';