+++ packages/ngx-localstorage/src/lib/providers/localstorage.provider.ts
@@ -1,4 +1,0 @@
-import { InjectionToken } from '@angular/core';
-import { ILocalStorageConfig } from '../types/localstorage.types';
-
-export const LOCALSTORAGE_CONFIG: InjectionToken<ILocalStorageConfig> = new InjectionToken<ILocalStorageConfig>('localstorageConfig');