+++ packages/ngx-localstorage/src/public-api.ts
@@ -1,6 +1,0 @@
-import { LocalstorageModule } from './lib/localstorage.module';
-import { LocalstorageService } from './lib/services/localstorage.service';
-
-export { LocalstorageModule, LocalstorageService };
-export { ILocalStorageConfig, IStorage, LocalStorageType } from './lib/types/localstorage.types';
-export { LOCALSTORAGE_CONFIG } from './lib/providers/localstorage.provider';