+++ packages/styleguide/src/aui/aui.module.ts
@@ -2,17 +2,13 @@ import { NgModule } from '@angular/core';
 import { FooterModule } from '../../../ngx-layout/src/lib/footer/public-api';
 import { LogoModule } from '../../../ngx-logo/src/public-api';
 import { HeaderModule } from '../../../ngx-layout/src/lib/header/public-api';
-import { NavigationMenuModule } from '../../../ngx-navigation-menu/src/public-api';
 import { CodeSnippetModule } from '../../../ngx-code-snippet/src/public-api';
-import { AnalyticsModule } from '../../../ngx-analytics/src/public-api';
-import { AgendaModule } from '../../../ngx-agenda/src/public-api';
 import { AvatarModule } from '../../../ngx-avatar/src/public-api';
 import { CalendarModule } from '../../../ngx-calendar/src/public-api';
 import { ContextModule } from '../../../ngx-context/src/public-api';
 import { FlyoutModule } from '../../../ngx-flyout/src/lib/flyout/flyout.module';
 import { FlyoutButtonModule } from '../../../ngx-flyout/src/lib/flyout-button/flyout-button.module';
 import { IconModule } from '../../../ngx-icon/src/public-api';
-import { LocalstorageModule } from '../../../ngx-localstorage/src/public-api';
 import { PaginationModule } from '../../../ngx-pagination/src/lib/pagination/public-api';
 import { ItemCounterModule } from '../../../ngx-pagination/src/lib/item-counter/public-api';
 import { ProgressBarModule } from '../../../ngx-progress-bar/src/public-api';
@@ -26,7 +22,7 @@ import { RangeSliderModule } from '../../../ngx-forms/src/lib/range-slider/publi
 import { SearchFilterModule } from '../../../ngx-forms/src/lib/search-filter/public-api';
 import { TimepickerModule } from '../../../ngx-forms/src/lib/timepicker/public-api';
 import { UploadModule } from '../../../ngx-forms/src/lib/upload/public-api';
-import { HeroModule, ModalModule, PaneModule, SidebarModule } from '../../../ngx-layout/src/public-api';
+import { ModalModule } from '../../../ngx-layout/src/public-api';
 import { FilterModule, LabelsModule, WindowModule } from '../../../ngx-utils/src/public-api';
 import { LeafletModule } from '../../../ngx-leaflet/src/public-api';
 import { NotificationsModule } from '../../../ngx-notifications/src/lib/ngx-notifications.module';
@@ -37,18 +33,12 @@ import { NotificationsModule } from '../../../ngx-notifications/src/lib/ngx-noti
     LogoModule,
     HeaderModule,
     CodeSnippetModule,
-    NavigationMenuModule.configure(),
-    AnalyticsModule,
-    AgendaModule,
     AvatarModule,
     CalendarModule,
     ContextModule,
     FlyoutModule,
     FlyoutButtonModule,
     IconModule,
-    LocalstorageModule.forRoot({
-      storageType: 'sessionStorage'
-    }),
     PaginationModule,
     ItemCounterModule,
     ProgressBarModule,
@@ -71,10 +61,7 @@ import { NotificationsModule } from '../../../ngx-notifications/src/lib/ngx-noti
     UploadModule,
     FooterModule,
     HeaderModule,
-    HeroModule,
     ModalModule,
-    PaneModule,
-    SidebarModule,
     FilterModule,
     LabelsModule,
     WindowModule,
@@ -85,17 +72,13 @@ import { NotificationsModule } from '../../../ngx-notifications/src/lib/ngx-noti
     FooterModule,
     CodeSnippetModule,
     LogoModule,
-    NavigationMenuModule,
     HeaderModule,
-    AnalyticsModule,
-    AgendaModule,
     AvatarModule,
     CalendarModule,
     ContextModule,
     FlyoutModule,
     FlyoutButtonModule,
     IconModule,
-    LocalstorageModule,
     PaginationModule,
     ItemCounterModule,
     ProgressBarModule,
@@ -111,10 +94,7 @@ import { NotificationsModule } from '../../../ngx-notifications/src/lib/ngx-noti
     UploadModule,
     FooterModule,
     HeaderModule,
-    HeroModule,
     ModalModule,
-    PaneModule,
-    SidebarModule,
     FilterModule,
     LabelsModule,
     WindowModule,