+++ packages/styleguide/src/examples/pages/aui-user-menu/aui-user-menu.page.ts
@@ -48,7 +48,7 @@ public loginUser(){
 
   public useComponentTagSnippet = `<aui-user-menu [user]="userObject" flyoutSize="Medium" [notificationsCount]="100"
 	(logout$)="logoutUser()" (login$)="loginUser()" [translations]="translations">
-	<div class="u-margin-lg u-text-center">
+	<div class="u-margin-l u-text-center">
 		Your Content Here
 	</div>
 </aui-user-menu>`;